[service] Move TracerProvider initialization to service/telemetry package
