+++ service/telemetry.go
@@ -5,7 +5,6 @@ package service // import "go.opentelemetry.io/collector/service"
 
 import (
 	"context"
-	"errors"
 	"net"
 	"net/http"
 	"strconv"
@@ -13,16 +12,10 @@ import (
 	ocmetric "go.opencensus.io/metric"
 	"go.opencensus.io/metric/metricproducer"
 	"go.opentelemetry.io/contrib/config"
-	"go.opentelemetry.io/contrib/propagators/b3"
-	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/metric"
 	noopmetric "go.opentelemetry.io/otel/metric/noop"
-	"go.opentelemetry.io/otel/propagation"
 	sdkmetric "go.opentelemetry.io/otel/sdk/metric"
 	"go.opentelemetry.io/otel/sdk/resource"
-	sdktrace "go.opentelemetry.io/otel/sdk/trace"
-	"go.opentelemetry.io/otel/trace"
-	nooptrace "go.opentelemetry.io/otel/trace/noop"
 	"go.uber.org/multierr"
 	"go.uber.org/zap"
 
@@ -34,20 +27,11 @@ import (
 const (
 	zapKeyTelemetryAddress = "address"
 	zapKeyTelemetryLevel   = "level"
-
-	// supported trace propagators
-	traceContextPropagator = "tracecontext"
-	b3Propagator           = "b3"
-)
-
-var (
-	errUnsupportedPropagator = errors.New("unsupported trace propagator")
 )
 
 type telemetryInitializer struct {
 	ocRegistry *ocmetric.Registry
 	mp         metric.MeterProvider
-	tp         trace.TracerProvider
 	servers    []*http.Server
 
 	disableHighCardinality bool
@@ -57,7 +41,6 @@ type telemetryInitializer struct {
 func newColTelemetry(disableHighCardinality bool, extendedConfig bool) *telemetryInitializer {
 	return &telemetryInitializer{
 		mp:                     noopmetric.NewMeterProvider(),
-		tp:                     nooptrace.NewTracerProvider(),
 		disableHighCardinality: disableHighCardinality,
 		extendedConfig:         extendedConfig,
 	}
@@ -74,32 +57,7 @@ func (tel *telemetryInitializer) init(res *resource.Resource, logger *zap.Logger
 	}
 
 	logger.Info("Setting up own telemetry...")
-
-	if tp, err := tel.initTraces(res, cfg); err == nil {
-		tel.tp = tp
-	} else {
-		return err
-	}
-
-	if tp, err := textMapPropagatorFromConfig(cfg.Traces.Propagators); err == nil {
-		otel.SetTextMapPropagator(tp)
-	} else {
-		return err
-	}
-
 	return tel.initMetrics(res, logger, cfg, asyncErrorChannel)
-}
-
-func (tel *telemetryInitializer) initTraces(res *resource.Resource, cfg telemetry.Config) (trace.TracerProvider, error) {
-	opts := []sdktrace.TracerProviderOption{}
-	for _, processor := range cfg.Traces.Processors {
-		sp, err := proctelemetry.InitSpanProcessor(context.Background(), processor)
-		if err != nil {
-			return nil, err
-		}
-		opts = append(opts, sdktrace.WithSpanProcessor(sp))
-	}
-	return proctelemetry.InitTracerProvider(res, opts)
 }
 
 func (tel *telemetryInitializer) initMetrics(res *resource.Resource, logger *zap.Logger, cfg telemetry.Config, asyncErrorChannel chan error) error {
@@ -170,19 +128,4 @@ func (tel *telemetryInitializer) shutdown() error {
 		}
 	}
 	return errs
-}
-
-func textMapPropagatorFromConfig(props []string) (propagation.TextMapPropagator, error) {
-	var textMapPropagators []propagation.TextMapPropagator
-	for _, prop := range props {
-		switch prop {
-		case traceContextPropagator:
-			textMapPropagators = append(textMapPropagators, propagation.TraceContext{})
-		case b3Propagator:
-			textMapPropagators = append(textMapPropagators, b3.New())
-		default:
-			return nil, errUnsupportedPropagator
-		}
-	}
-	return propagation.NewCompositeTextMapPropagator(textMapPropagators...), nil
 }
